requirements.txt
termcolor
paho-mqtt~=1.6.1
requests